Removes the Windows v1 runtime
