+++ cmd/containerd/builtins_windows.go
@@ -1,4 +1,4 @@
-// +build !windows_v2
+// +build windows
 
 /*
    Copyright The containerd Authors.
@@ -21,7 +21,7 @@ package main
 import (
 	_ "github.com/containerd/containerd/diff/lcow"
 	_ "github.com/containerd/containerd/diff/windows"
+	_ "github.com/containerd/containerd/runtime/v2"
 	_ "github.com/containerd/containerd/snapshots/lcow"
 	_ "github.com/containerd/containerd/snapshots/windows"
-	_ "github.com/containerd/containerd/windows"
 )
