+++ services/tasks/local_windows.go
@@ -16,48 +16,20 @@
    limitations under the License.
 */
 
-package windows
+package tasks
 
 import (
-	"errors"
-	"sync"
+	"github.com/containerd/containerd/plugin"
+	"github.com/containerd/containerd/runtime"
 )
 
-type pidPool struct {
-	sync.Mutex
-	pool map[uint32]struct{}
-	cur  uint32
+var tasksServiceRequires = []plugin.Type{
+	plugin.RuntimePluginV2,
+	plugin.MetadataPlugin,
+	plugin.TaskMonitorPlugin,
 }
 
-func newPidPool() *pidPool {
-	return &pidPool{
-		pool: make(map[uint32]struct{}),
-	}
-}
-
-func (p *pidPool) Get() (uint32, error) {
-	p.Lock()
-	defer p.Unlock()
-
-	pid := p.cur + 1
-	for pid != p.cur {
-		// 0 is reserved and invalid
-		if pid == 0 {
-			pid = 1
-		}
-		if _, ok := p.pool[pid]; !ok {
-			p.cur = pid
-			p.pool[pid] = struct{}{}
-			return pid, nil
-		}
-		pid++
-	}
-
-	return 0, errors.New("pid pool exhausted")
-}
-
-func (p *pidPool) Put(pid uint32) {
-	p.Lock()
-	delete(p.pool, pid)
-	p.Unlock()
+// loadV1Runtimes on Windows there are no v1 runtimes
+func loadV1Runtimes(ic *plugin.InitContext) (map[string]runtime.PlatformRuntime, error) {
+	return make(map[string]runtime.PlatformRuntime), nil
 }
